MP OpenAPI 4.0: Implement document merging
Extend our code which merges OpenAPI documents so that it can handle the
OpenAPI 3.1 documents produced by MP OpenAPI 4.0.

Merging is done using the MP OpenAPI model API, so we need to create
abstractions to cope with the cases where the API has breaking changes.
