+++ dev/io.openliberty.microprofile.openapi.2.0.internal/src/io/openliberty/microprofile/openapi20/internal/merge/NameType.java
@@ -4,7 +4,7 @@
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -23,11 +23,13 @@ public enum NameType {
     HEADERS("headers"),
     LINKS("links"),
     PARAMETERS("parameters"),
+    PATH_ITEMS("pathItems"),
     REQUEST_BODIES("requestBodies"),
     RESPONSES("responses"),
     SCHEMAS("schemas"),
     SECURITY_SCHEMES("securitySchemes"),
-    PATHS(null);
+    PATHS(null),
+    WEBHOOKS(null);
 
     private final String referencePrefix;
