+++ dev/io.openliberty.microprofile.openapi.3.1.internal.services/src/io/openliberty/microprofile/openapi31/internal/services/impl/OpenAPI30ModelWalker.java
@@ -1,0 +1,21 @@
+/*******************************************************************************
+ * Copyright (c) 2024 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package io.openliberty.microprofile.openapi31.internal.services.impl;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ConfigurationPolicy;
+
+import io.openliberty.microprofile.openapi20.internal.utils.OpenAPIModelWalker;
+import io.openliberty.microprofile.openapi20.internal.utils.OpenAPIModelWalkerImpl;
+
+@Component(service = OpenAPIModelWalker.class, configurationPolicy = ConfigurationPolicy.IGNORE)
+public class OpenAPI30ModelWalker extends OpenAPIModelWalkerImpl {
+
+}
