+++ dev/io.openliberty.microprofile.openapi.4.0.internal.services/bnd.bnd
@@ -54,4 +54,9 @@ WS-TraceGroup: MPOPENAPI
     com.ibm.ws.org.slf4j.api;version=latest, \
     com.ibm.ws.org.slf4j.jdk14;version=latest, \
     io.openliberty.org.eclipse.microprofile.config.3.1, \
-    org.jboss.logging:jboss-logging
+    org.jboss.logging:jboss-logging,\
+    io.openliberty.microprofile.openapi.2.0.internal_test;version=latest,\
+    io.openliberty.com.fasterxml.jackson;version=latest,\
+    com.ibm.ws.kernel.service;version=latest
+
+
