+++ dev/io.openliberty.microprofile.openapi.4.0.internal.services/src/io/openliberty/microprofile/openapi40/internal/services/impl/OpenAPI31MergeProcessor.java
@@ -4,26 +4,57 @@
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 package io.openliberty.microprofile.openapi40.internal.services.impl;
 
 import java.util.List;
+import java.util.Map;
 
+import org.eclipse.microprofile.openapi.models.Components;
+import org.eclipse.microprofile.openapi.models.OpenAPI;
+import org.eclipse.microprofile.openapi.models.PathItem;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
 
+import io.openliberty.microprofile.openapi20.internal.merge.MergeProcessorImpl;
+import io.openliberty.microprofile.openapi20.internal.merge.NameProcessor.DocumentNameProcessor;
+import io.openliberty.microprofile.openapi20.internal.merge.NameType;
 import io.openliberty.microprofile.openapi20.internal.services.MergeProcessor;
 import io.openliberty.microprofile.openapi20.internal.services.OpenAPIProvider;
 
-@Component(service = MergeProcessor.class, configurationPolicy = ConfigurationPolicy.IGNORE)
-public class OpenAPI31MergeProcessor implements MergeProcessor {
+@Component(configurationPolicy = ConfigurationPolicy.IGNORE)
+public class OpenAPI31MergeProcessor extends MergeProcessorImpl implements MergeProcessor {
 
     @Override
-    public OpenAPIProvider mergeDocuments(List<OpenAPIProvider> documents) {
-        // TODO: implement merging
-        return documents.get(0);
+    protected MergeProcessorInstance createMergeProcessorInstance(List<OpenAPIProvider> documents) {
+        return new MergeProcessorInstance31(documents);
+    }
+
+    protected class MergeProcessorInstance31 extends MergeProcessorInstance {
+
+        public MergeProcessorInstance31(List<OpenAPIProvider> providers) {
+            super(providers);
+        }
+
+        @Override
+        protected void renameClashingComponents(OpenAPI document, DocumentNameProcessor documentNameProcessor) {
+            super.renameClashingComponents(document, documentNameProcessor);
+
+            Components components = document.getComponents();
+            if (components != null) {
+                components.setPathItems(renameComponents(NameType.PATH_ITEMS, components.getPathItems(), documentNameProcessor));
+            }
+        }
+
+        @Override
+        protected void renameClashingWebhooks(OpenAPI document, DocumentNameProcessor documentNameProcessor) {
+            Map<String, PathItem> webhooks = document.getWebhooks();
+            if (webhooks != null) {
+                document.setWebhooks(renameComponents(NameType.WEBHOOKS, webhooks, documentNameProcessor));
+            }
+        }
     }
 
 }
